refactor: split the workbench view provider into focused modules
The 1048-line provider now owns only the webview lifecycle and publish
queue (~300 lines). The HTML document moves to workbench/view-html.ts
(template unchanged), the webview message routing to
workbench/view-messages.ts with an explicit message context, and the
payload parsers/validators to workbench/input-validators.ts. The
responsive-layout test follows the template to its new home.
